+++ R/outputs.R
@@ -74,7 +74,7 @@ add_stats <- function(df, mode = NULL, sample = NULL, prefix = ""){
 
 
 extract_indicators <- function(naomi_fit, naomi_mf) {
-
+  
   get_est <- function(varname,
                       indicator_id,
                       calendar_quarter,
@@ -155,41 +155,57 @@ extract_art_attendance <- function(naomi_fit, naomi_mf) {
     m_artattend_ij_t1 <- mode$artattend_ij_t1_out
     m_artnum_reside_t1 <- mode$artnum_t1_out[v$reside_out_idx]
     m_artnum_attend_t1 <- mode$artattend_t1_out[v$attend_out_idx]
-    m_prop_residents_t1 <- m_artattend_ij_t1 / m_artnum_reside_t1
-    m_prop_attendees_t1 <- m_artattend_ij_t1 / m_artnum_attend_t1
     
     m_artattend_ij_t2 <- mode$artattend_ij_t2_out
     m_artnum_reside_t2 <- mode$artnum_t2_out[v$reside_out_idx]
     m_artnum_attend_t2 <- mode$artattend_t2_out[v$attend_out_idx]
+  } else {
+    m_artattend_ij_t1 <- NULL
+    m_artattend_ij_t2 <- NULL
+  }
+
+  if(!is.null(m_artattend_ij_t1)) {
+    m_prop_residents_t1 <- m_artattend_ij_t1 / m_artnum_reside_t1
+    m_prop_attendees_t1 <- m_artattend_ij_t1 / m_artnum_attend_t1
+  } else {
+    m_prop_residents_t1 <- NULL
+    m_prop_attendees_t1 <- NULL
+  }
+
+  if(!is.null(m_artattend_ij_t2)) {
     m_prop_residents_t2 <- m_artattend_ij_t2 / m_artnum_reside_t2
     m_prop_attendees_t2 <- m_artattend_ij_t2 / m_artnum_attend_t2
   } else {
-    m_artattend_ij_t1 <- NULL
-    m_prop_residents_t1 <- NULL
-    m_prop_attendees_t1 <- NULL
-    m_artattend_ij_t2 <- NULL
     m_prop_residents_t2 <- NULL
     m_prop_attendees_t2 <- NULL
   }
-  
+
   if(!is.null(naomi_fit$sample)) {
 
     s_artattend_ij_t1 <- naomi_fit$sample$artattend_ij_t1_out
     s_artnum_reside_t1 <- naomi_fit$sample$artnum_t1_out[v$reside_out_idx, ]
     s_artnum_attend_t1 <- naomi_fit$sample$artattend_t1_out[v$attend_out_idx, ]
-    s_prop_residents_t1 <- s_artattend_ij_t1 / s_artnum_reside_t1
-    s_prop_attendees_t1 <- s_artattend_ij_t1 / s_artnum_attend_t1
    
     s_artattend_ij_t2 <- naomi_fit$sample$artattend_ij_t2_out
     s_artnum_reside_t2 <- naomi_fit$sample$artnum_t2_out[v$reside_out_idx, ]
     s_artnum_attend_t2 <- naomi_fit$sample$artattend_t2_out[v$attend_out_idx, ]
+  } else {
+    s_artattend_ij_t1 <- NULL
+    s_artattend_ij_t2 <- NULL
+  }
+
+  if(!is.null(s_artattend_ij_t1)) {
+    s_prop_residents_t1 <- s_artattend_ij_t1 / s_artnum_reside_t1
+    s_prop_attendees_t1 <- s_artattend_ij_t1 / s_artnum_attend_t1
+  } else {
+    s_prop_residents_t1 <- NULL
+    s_prop_attendees_t1 <- NULL
+  }
+
+  if(!is.null(s_artattend_ij_t2)) {
     s_prop_residents_t2 <- s_artattend_ij_t2 / s_artnum_reside_t2
     s_prop_attendees_t2 <- s_artattend_ij_t2 / s_artnum_attend_t2
   } else {
-    s_artattend_ij_t1 <- NULL
-    s_prop_residents_t1 <- NULL
-    s_prop_attendees_t1 <- NULL
-    s_artattend_ij_t2 <- NULL
     s_prop_residents_t2 <- NULL
     s_prop_attendees_t2 <- NULL
   }